+++ src/test/java/model/request/RequestLineTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.params.provider.ValueSource;
 import webserver.http.Method;
 import webserver.http.Type;
 import webserver.http.Version;
+import webserver.http.request.RequestLine;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.entry;
